SIMKDOWN,EDC: Fix potential bug with reading in encrypted arguments
The type of encryptedArgument and encryptedArgument2 is int64_t, but we
were reading them as int - read them as int64_t instead.

Note that none of the testcases generated so far actually triggered this
bug - since the maxDocLength is <= 1e9, all of the encryptedArgument/
encryptedArgument2's are approximately this size (maybe as large as the
next power of 2, minus 1), and so fit into an int.  With a significantly
larger maxDocLength, though, this would almost certainly trigger a bug.

+++ my-own-problems/simple-markdown-editor/simple-markdown-editor-editorial.cpp
@@ -461,7 +461,7 @@ int main()
         for (auto& query : queries)
         {
             const auto queryType = read<char>();
-            query.encryptedArgument = read<int>();
+            query.encryptedArgument = read<int64_t>();
             switch (queryType)
             {
                 case 'F':
@@ -469,7 +469,7 @@ int main()
                     break;
                 case 'N':
                     query.type = Query::InsertNonFormatting;
-                    query.encryptedArgument2 = read<int>();
+                    query.encryptedArgument2 = read<int64_t>();
                     break;
                 case 'Q':
                     query.type = Query::IsRangeFormatted;
